qcom-cpufreq: Boost all online CPUs when exiting suspend
In order to improve resume performance when the CPU is waking up from
suspend (i.e., deep sleep), boost all online CPUs to their respective
maximum frequencies for 20 ms. The power cost of boosting everything to
the max for 20 ms is negligible, so this doesn't significantly increase
power consumption.

The boost is kicked off right before all the userspace processes are
thawed, as that's when the boost is needed most.

Change-Id: Ie3c63b4e885d4cc0b961089e697109f5fa0c0abd
Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>
Signed-off-by: Davide Garberi <dade.garberi@gmail.com>
Signed-off-by: DennySPB <dennyspb@gmail.com>